requirements.txt
discord.py
requests
python-dotenv
beautifulsoup4
pytz
aiohttp
